enhance: better error report

+++ boltconn/src/config/proxy_provider.rs
@@ -45,41 +45,49 @@ pub async fn read_proxy_schema(
 ) -> anyhow::Result<HashMap<String, ProxySchema>> {
     let mut table = HashMap::new();
     // concurrently download rules
-    let tasks: Vec<JoinHandle<anyhow::Result<(String, ProxySchema)>>> = providers
+    let tasks: HashMap<String, JoinHandle<anyhow::Result<ProxySchema>>> = providers
         .clone()
         .into_iter()
         .map(|(name, item)| {
             let root_path = config_path.to_path_buf();
-            tokio::spawn(async move {
-                match item {
-                    ProxyProvider::File { path } => {
-                        let content: ProxySchema = serde_yaml::from_str(
-                            fs::read_to_string(safe_join_path(&root_path, &path)?)?.as_str(),
-                        )?;
-                        Ok((name.clone(), content))
+            (
+                name,
+                tokio::spawn(async move {
+                    match item {
+                        ProxyProvider::File { path } => {
+                            let content: ProxySchema = serde_yaml::from_str(
+                                fs::read_to_string(safe_join_path(&root_path, &path)?)?.as_str(),
+                            )?;
+                            Ok(content)
+                        }
+                        ProxyProvider::Http { url, path, .. } => {
+                            let full_path = safe_join_path(&root_path, &path)?;
+                            let content: ProxySchema =
+                                if !force_update && full_path.as_path().exists() {
+                                    serde_yaml::from_str(
+                                        fs::read_to_string(full_path.as_path())?.as_str(),
+                                    )?
+                                } else {
+                                    let resp = reqwest::get(url).await?;
+                                    let text = resp.text().await?;
+                                    let content: ProxySchema = serde_yaml::from_str(text.as_str())?;
+                                    // security: `full_path` should be (layers of) subdir of `root_path`,
+                                    //           so arbitrary write should not happen
+                                    fs::write(full_path.as_path(), text)?;
+                                    content
+                                };
+                            Ok(content)
+                        }
                     }
-                    ProxyProvider::Http { url, path, .. } => {
-                        let full_path = safe_join_path(&root_path, &path)?;
-                        let content: ProxySchema = if !force_update && full_path.as_path().exists()
-                        {
-                            serde_yaml::from_str(fs::read_to_string(full_path.as_path())?.as_str())?
-                        } else {
-                            let resp = reqwest::get(url).await?;
-                            let text = resp.text().await?;
-                            let content: ProxySchema = serde_yaml::from_str(text.as_str())?;
-                            // security: `full_path` should be (layers of) subdir of `root_path`,
-                            //           so arbitrary write should not happen
-                            fs::write(full_path.as_path(), text)?;
-                            content
-                        };
-                        Ok((name.clone(), content))
-                    }
-                }
-            })
+                }),
+            )
         })
         .collect();
-    for task in tasks {
-        let (name, content) = task.await??;
+    for (name, task) in tasks {
+        let content = match task.await? {
+            Ok(c) => c,
+            Err(e) => return Err(anyhow::anyhow!("In proxy provider {}: {}", name, e)),
+        };
         table.insert(name, content);
     }
     Ok(table)
